app/src/requirements.txt
Django==2.2.10
django-environ==0.4.4
celery==4.1.1
flower==0.9.2
psycopg2-binary==2.8.3
redis>=3.2.0
django-extensions==2.0.6
sentry-sdk==0.4.1
tornado<6
django-constance==2.6.0